+++ src/func_0200af20.c
@@ -77,5 +77,5 @@ void func_0200af20(char *c, struct Vector3 *v1, struct Vector3 *v2, short *out)
 
     *out = 0xe38;
     _ZN7Clipper13Func_020156DCEv(data_0209f43c, *(int *)(c + 0xf8), *out, *(int *)(c + 0xfc), *(int *)(c + 0x100));
-    *(int *)(int)(((long long)(int)(c + 0x154))) &= ~1;
+    *(int *)(int)(c + 0x154) &= ~1;
 }
