+++ src/func_0204605c.c
@@ -4,7 +4,7 @@ struct Obj { void *pad0; struct Element *arr; };
 void func_0204605c(struct Obj *self, unsigned int value, int index)
 {
     struct Element *e = &self->arr[index];
-    unsigned int *flags = (unsigned int *)(((long long)(int)&e->flags));
+    unsigned int *flags = (unsigned int *)(&e->flags);
     *flags &= ~0x3f000000;
     *flags |= value << 24;
 }
