+++ src/func_ov002_020b2a34.c
@@ -9,7 +9,7 @@ void func_ov002_020b2a34(char* c, char* p)
 {
     int state = *(int*)(c + 0x3a0);
     if (state == 1) {
-        *(int*)(((long long)(int)(c + 0x60))) += 0x50000;
+        *(int*)(c + 0x60) += 0x50000;
         func_ov002_020b16c4(c, p);
         return;
     }
